revamp fire pressureplate logic
previously the logic of sudden movement where the position of the button
is teleported is bound to cause issues. even my previous fix was not
applicable universally because the race condition between the button
being pressed down and the player falling down vs ontriggerexit still
exists.
I have added a new logic that make it so the button is slowly going up
or down, so the character wont experience a sudden fall which makes
their hitbox exit the button's hitbox.